writer: setTimestamp taking ulong

+++ modules/signal/include/dls2/signal/signal_writer.hpp
@@ -30,7 +30,7 @@ namespace dls
 		@brief Set the timestamp of the signal stored in any child class
 		@param[in] timestamp		
 		*/
-		virtual void setTimestamp(double timestamp) override;
+		virtual void setTimestamp(unsigned long long timestamp) override;
 
 		/*! @brief*/
 		void setSignal(const SignalType&);
